# docs/swagger.yaml
openapi: 3.0.3
info:
  title: BotBuilder API
  description: |
    AI-Powered Chatbot Builder Platform API

    ## Overview
    BotBuilder provides a comprehensive API for building, deploying, and managing
    intelligent chatbots across multiple platforms.

    ## Authentication
    Most endpoints require JWT authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-token>
    ```

    ## Rate Limiting
    - General API: 100 requests per 15 minutes
    - Authentication: 5 requests per 15 minutes
    - AI Endpoints: 20 requests per minute

    ## Error Responses
    All errors follow a consistent format:
    ```json
    {
      "success": false,
      "error": "Error message",
      "code": "ERROR_CODE"
    }
    ```

  version: 2.0.0
  contact:
    name: BotBuilder Support
    email: support@botbuilder.com
    url: https://github.com/your-org/BotBuilder
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000/api
    description: Development server
  - url: https://api.your-domain.com/api
    description: Production server
...
paths:
...
  /knowledge/{kbId}/search:
    post:
      tags:
        - Knowledge Base
      summary: Search knowledge base
      description: Semantic search using embeddings
      operationId: searchKnowledgeBase
      security:
        - bearerAuth: []
      parameters:
        - name: kbId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - query
              properties:
                query:
                  type: string
                limit:
                  type: integer
                  default: 5
                threshold:
                  type: number
                  default: 0.7
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  results:
                    type: array
                    items:
                      type: object
                      properties:
                        content:
                          type: string
                        document_name:
                          type: string
                        similarity:
                          type: number

  # ========================================
  # ANALYTICS
  # ========================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token authentication
...
security:
  - bearerAuth: []